pyproject.toml
[tool.poetry]
name = "ecommerce-dicria-backend-mrlomnds"
version = "0.1.0"
description = "API para esports"
authors = ["MrloMnds <murilol147@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
ipdb = "^0.13.13"
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
